New feature : Aggregated packet sub-topic selection
This new feature allows the user to select the aggregated
sub-topics names for archive and loop packets if needed, this allows to
subscribe to two different topics when both archive and loop bindings
are enabled.
The defaults are to still continue to post on the "loop" topic for both
types in order to not break existing users applications.

Added a bit of documentation for paho-mqtt installation and for the two
new options.