Reusable MQTT client
Establishing a new MQTT client for each loop iteration is quite wasteful,
especially when TLS is in use. This change establishes and uses a single
client. Automatic reconnection is handled by Paho-MQTT.

For each call to process_record(), we'll check to see if an MQTT client
has already been established. If so, we'll use it. If not, and it's been
at least retry_wait seconds since we last tried, we'll try to establish
one. socket exceptions (expected causes of failure) are caught and logged.
Other exceptions are raised up. If a client is not available to do the
requested publish, a weewx.restx.FailedPost exception is raised.

resolves #9